+++ src/main/java/com/example/Tozin_Solutions_back_end/dto/pecaDTO/AtualizarPecaDTO.java
@@ -2,9 +2,9 @@ package com.example.Tozin_Solutions_back_end.dto.pecaDTO;
 
 public class AtualizarPecaDTO {
     private String nome;
-    private Integer quantidadeMinima;
+    private Double quantidadeMinima;
 
-    public AtualizarPecaDTO(String nome, Integer quantidadeMinima) {
+    public AtualizarPecaDTO(String nome, Double quantidadeMinima) {
         this.nome = nome;
         this.quantidadeMinima = quantidadeMinima;
     }
@@ -13,7 +13,7 @@ public class AtualizarPecaDTO {
         return nome;
     }
 
-    public Integer getQuantidadeMinima() {
+    public Double getQuantidadeMinima() {
         return quantidadeMinima;
     }
 }
